pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data_quality_analysis"
version = "0.0.1"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "GitPython>=3.1.37",
  "botok >= 0.8.1",
  "PyGithub >= 2.1.1"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]


[project.urls]
"Homepage" = "https://github.com/OpenPecha/openpecha-project-template"
"Bug Tracker" = "https://github.com/OpenPecha/openpecha-project-template/issues"
